// front_end/frontend_serv_vaccine/src/interfaces/relationship-interfaces.tsx
// Generated by https://quicktype.io

export interface RelationShipResponse {
    desde:         number;
    limite:        number;
    relationships: Relationship[];
    total:         number;
}

export interface Relationship {
    _id:    ID;
    name:   string;
    status: boolean;
}

export interface ID {
    $oid: string;
}